combine python 2 and python 3

+++ Maya/tools/apiWeights.py
@@ -176,7 +176,7 @@ class ApiWeights():
             amountInfluences = len(infIndices)
             weights = skinFn.getWeights( meshPath , vertexComp)
             _wght = weights[0]
-            weights = [[_wght[i+j*amountInfluences] for i in range(amountInfluences)] for j in range(len(_wght)/amountInfluences)]
+            weights = [[_wght[int(i+j*amountInfluences)] for i in range(amountInfluences)] for j in range(int(len(_wght)/amountInfluences))]
             self.meshWeights[node] = weights
             
             jointInfluences = [infDags[x].fullPathName() for x in range(len(infIndices))]
